requirements.txt
python-dotenv>=1.2.2
python-telegram-bot>=22.7
requests>=2.33.1
PyPDF2>=3.0.0